+++ tests/one_shot.v
@@ -90,9 +90,9 @@ Proof.
       subst; wp_match; [done|].
     wp_bind (! _)%E.
     iInv N as "[[Hl >Hγ]|H]"; last iDestruct "H" as (m') "[Hl Hγ]".
-    { by iDestruct (own_valid_2 with "Hγ Hγ'") as %?. }
+    { by iCombine "Hγ Hγ'" gives %?. }
     wp_load. Show.
-    iDestruct (own_valid_2 with "Hγ Hγ'") as %?%to_agree_op_inv_L; subst.
+    iCombine "Hγ Hγ'" gives %?%to_agree_op_inv_L; subst.
     iModIntro. iSplitL "Hl".
     { iNext; iRight; by eauto. }
     wp_smart_apply wp_assert. wp_pures. by case_bool_decide.